requirements.txt
pmdarima==1.5.3
statsmodels==0.11.1
streamlit==0.57.3
numpy==1.16.5
panda==0.3.1
plotly==4.6.0
lxml==4.4.1
